issue: Float format for storing relationship end_date
The Relationship class reads relationship data from the backend and converts the end date to float format. However this leads to loss of precision. The mentorship relation will then expire some time before or after the displayed date. This also creates problems for sorting or filtering mentorship requests.
For example the sever returns "1585679404" but this is getting rounded off to 1.5795927E9 = 1579592704

To fix this either double or a special date class should be used for storing dates.